Added initial shader assets (#36)
I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself