35 crafting objects shader (#41)
* Added initial shader assets

I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself

* Added shaders for items

I have added a shader to indicate the rarity of the crafting items using a combination of fresnel shader and a noise texture to give items a polution infected style