Game audio and shaders (#44)
* Added Wwise integration (#33)

* Integration of the Wwise project

Added the folder containing the Wwise project for the game

* Added the AkAudio module to the project

* Game music (#34)

* Integration of the Wwise project

Added the folder containing the Wwise project for the game

* Added the AkAudio module to the project

* Added music for the game and some UI SoundFX

- I have added background music for the title screen as well as for the main game.
- I have added sound effects when the user hovers/selects a menu option.

* Added missing Wwise assets

Some assets were missing because they were ignored in the .gitignore file

* Fix MyCharacter.cpp file

The file still had a function that was removed and prevented me from building the game after my merge. It is weird because I did not have any issue running the game before checking out to the merge branch (dev-audio)

* Added additional SoundFX to the game

I have added sound effects for the player death as well as when the player picks up crafting objects

* Added initial shader assets (#36)

I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself

* Corrected the game's audio

Because a lot of the code for the audio was uassets, the game audio was lost when merging changes from another branch

* 35 crafting objects shader (#41)

* Added initial shader assets

I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself

* Added shaders for items

I have added a shader to indicate the rarity of the crafting items using a combination of fresnel shader and a noise texture to give items a polution infected style

* Added the audio files

I have added the audio files since it's the end of the project

* Added Wwise integration (#33)

* Integration of the Wwise project

Added the folder containing the Wwise project for the game

* Added the AkAudio module to the project

* Game music (#34)

* Integration of the Wwise project

Added the folder containing the Wwise project for the game

* Added the AkAudio module to the project

* Added music for the game and some UI SoundFX

- I have added background music for the title screen as well as for the main game.
- I have added sound effects when the user hovers/selects a menu option.

* Added missing Wwise assets

Some assets were missing because they were ignored in the .gitignore file

* Added additional SoundFX to the game

I have added sound effects for the player death as well as when the player picks up crafting objects

* Added initial shader assets (#36)

I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself

* Corrected the game's audio

Because a lot of the code for the audio was uassets, the game audio was lost when merging changes from another branch

* 35 crafting objects shader (#41)

* Added initial shader assets

I have added the initial shader assets to give crafting items a unique look depending on their rarity. This is still a rough version of the shaders since the layer concept is not working. Ideally, we would like the shader to layer itself on top of the objects textures instead of being the texture itself

* Added shaders for items

I have added a shader to indicate the rarity of the crafting items using a combination of fresnel shader and a noise texture to give items a polution infected style

* Added the audio files

I have added the audio files since it's the end of the project

* Added back the audio in the blueprint when it wasn't

* Fixed some rebase prbl, removed some unused asset, adapted the level to play the sound

* Added additional audio features and fixed/corrected various issues

- Added more dynamic music to the game with 3 different main music states and parts. The main music changes depending on if the player is safe, getting hurt, or almost dying.
- Cleaned some console logs that were not necessary as well as changed the text of some log errors that were not correct.
- Disabled nanite since it was not being used in our game and also caused unwanted warnings in the logs as well as an error. This involved removing nanite on static meshes, disabling it in the engine, and changing the shadow map settings for the project.
- Fix the UI navigation in standalone/build, which was implemented but somehow got removed along the way.

* Fixed rebases conflicts and made a victory menu

---------

Co-authored-by: MohETS <72469648+MohETS@users.noreply.github.com>
Co-authored-by: MohETS <mohmantis@gmail.com>